+++ core/include/Contour.h
@@ -25,7 +25,7 @@ class Contour {
   void magneticBoundaries(const TH2F* hCL);
   inline void setSigma(int s) { m_sigma = s; };
   void setStyle(int linecolor, int linestyle, int linewidth, int fillcolor, int fillstyle);
-  void setTransparency(float percent);
+  void setTransparency(double percent);
 
  private:
   TGraph* changePointOrder(TGraph* g, int pointId);
@@ -44,7 +44,7 @@ class Contour {
   int m_fillcolor;
   int m_fillstyle;
   int m_linewidth;
-  float m_alpha;
+  double m_alpha;
 };
 
 #endif
